add support for environment variables
a rule can specify the field "set":

"set" : ["VARIABLE",  value]

and if this rule is applied, the variable VARIABLE (may be lower-cast)
is remembered with value "value". This variable survives
between phases of instrumentation, so later phases of intrumentation
can use the variable in condition:

"condition" : ["VARIABLE", "desired_value"]

to check whether a variable has the desired value.
We use it now to check whether we instrumented some heap memory
allocation and if not, we do not add a check for memory leaks.

Restriction: does not work for global variables atm.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>